refactor by moving all files to authentication module
