+++ src/main/java/com/security/jdbc/SpringSecurityJdbcApplication.java
@@ -1,10 +1,7 @@
 package com.security.jdbc;
 
-import com.security.jdbc.repository.UserRepository;
-import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
-import org.springframework.stereotype.Component;
 
 @SpringBootApplication
 public class SpringSecurityJdbcApplication {
