+++ src/main/java/com/security/jdbc/authentication/repository/UserRepositoryRunner.java
@@ -1,4 +1,4 @@
-package com.security.jdbc.repository;
+package com.security.jdbc.authentication.repository;
 
 import org.springframework.boot.CommandLineRunner;
 import org.springframework.stereotype.Component;
@@ -15,7 +15,7 @@ public class UserRepositoryRunner implements CommandLineRunner {
     @Override
     public void run(String... args) throws Exception {
 
-        System.out.println("USER RUNNER: " + userRepository.findOneUser("rainhard@email.com"));
+        System.out.println("USER RUNNER: " + userRepository.findOneUserByEnabledAndEmail("rainhard@email.com"));
         System.out.println("AUTHORITY RUNNER: " + userRepository.getUserAuthoritiesByUserEmail("rainhard@email.com"));
 
     }
