+++ src/main/java/com/security/jdbc/authentication/security/jwt/JwtService.java
@@ -1,4 +1,4 @@
-package com.security.jdbc.security.jwt;
+package com.security.jdbc.authentication.security.jwt;
 
 import io.jsonwebtoken.Claims;
